Added Email services and manager approve controller

+++ java_ca/src/main/java/edu/nus/java_ca/service/LeaveServiceImpl.java
@@ -95,7 +95,7 @@ public class LeaveServiceImpl implements LeaveService{
 	}
 	
 	@Transactional
-	public Leave findLeaveById(Integer id) {
+	public Leave findLeaveById(Long id) {
 		return lrepo.findById(id).get();
 	}
